tidied up files slightly and can now append mcmc runs instead of overwriting
